Split out ResultItemValueProperties to its own file
Also, making the ResultValueProperty return the first item if no string is provided
